Update meta dictionary
Still, the following samples were not submitted:

/GluGluToRadionToHHTo2B2VTo2L2Nu_M-260_narrow_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM
/GluGluToRadionToHHTo2B2Tau_M-800_narrow_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM
/GluGluToRadionToHHTo2B2Tau_M-1750_narrow_13TeV-madgraph_correctedcfg/RunIIFall17MiniAODv2-PU2017_12Apr2018_94X_mc2017_realistic_v14-v1/MINIAODSIM

I was able to submit jobs for the 2nd and the 3rd datasets, but not for
the 1st dataset, b/c apparently it's available only on tiers that are on
the global CRAB blacklist. Submitted transfer request to T2_EE_Estonia
for this sample. No idea why the submission of 2nd and 3rd sample failed
earlier.