Remove signal_ggf_spin2_300_hh_bbtt sample from 2017
Even though the CRAB task has finished successully, only 4.1% of the events have been processed. The probable cause is that the dataset was still in "production" and only 4.1% of the events were available at the time of the submission.
Will have to resubmit the CRAB task again now that the dataset has been marked as "valid".