Use AnalysisConfig getters where possible
Decided not to create AnalysisConfig::isMC_TT() to initialize 'isMC_TT' because the latter is initialized to true if the process string contains 'TT' substring which means that it's set to true also for eg TTW and TTZ categories. Creating a similar function in AnalysisConfig would cause more confusion than needed.