Properly handle the case if data-driven DY bkg estimation is disabled
In the submission script, made the DY bkg options configurable via command line, so that when generating cfg files for eg signal sample only, the FW wouldn't abort because the signal samples are excluded from this procedure. Also, telling the FW to run the procedure using 'compWeights' is more flexible than editing some lines in the submission script.
In analyzeConfig_hh_bb2l.py, the shape uncertainties that are associated with the data-driven DY bakcground were always run, even if the FW was not configured to run MC closure jobs or even if the data-driven DY background estimation was disabled altogether.